# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Convoy API",
    "description": "\nConvoy is a truck discovery and availability marketplace API.\n\n## Authentication\n1. **POST /api/auth/send-otp** then **POST /api/auth/verify-otp** to log in.\n2. Use the returned **`accessToken`** as a Bearer JWT on all other endpoints:\n   `Authorization: Bearer <accessToken>`\n3. Path `user_id` must match the logged-in user (from the token).\n\n## Interactive docs\n- **Swagger UI:** `/docs`\n- **ReDoc:** `/redoc`\n- **OpenAPI JSON:** `/openapi.json`\n\n## Base path\nAll routes are served under **`/api`**.\n",
    "contact": {
      "name": "Convoy Support",
      "url": "https://github.com/SkandTrizip/convoy_prod"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Current environment"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/posts/create/{user_id}": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "Create Truck Post",
        "description": "Create a truck availability post (auto-expires after 24 hours).",
        "operationId": "create_truck_post_api_posts_create__user_id__post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTruckPostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/my-posts/{user_id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get My Posts",
        "description": "Get user's truck posts (active or expired).",
        "operationId": "get_my_posts_api_posts_my_posts__user_id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/reactivate/{post_id}": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "Reactivate Post",
        "description": "Reactivate an expired post for another 24 hours.",
        "operationId": "reactivate_post_api_posts_reactivate__post_id__post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/delete/{post_id}": {
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "Delete Post",
        "description": "Delete a truck post",
        "operationId": "delete_post_api_posts_delete__post_id__delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/edit/{post_id}": {
      "put": {
        "tags": [
          "posts"
        ],
        "summary": "Edit Post",
        "description": "Edit a truck post (vehicle, destination, current location) and reactivate/extend expiry.",
        "operationId": "edit_post_api_posts_edit__post_id__put",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTruckPostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTruckPostRequest": {
        "properties": {
          "vehicleId": {
            "type": "string",
            "title": "Vehicleid",
            "description": "UUID of a verified vehicle"
          },
          "origin": {
            "$ref": "#/components/schemas/Location"
          },
          "destinations": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Destinations",
            "description": "1-5 destinations for this route"
          },
          "currentLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "contactName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contactname",
            "description": "Contact name for this post. Defaults to the user's own name if omitted."
          },
          "contactNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contactnumber",
            "description": "Contact number for this post. Defaults to the user's own mobile if omitted."
          }
        },
        "type": "object",
        "required": [
          "vehicleId",
          "origin",
          "destinations",
          "currentLocation"
        ],
        "title": "CreateTruckPostRequest"
      },
      "EditTruckPostRequest": {
        "properties": {
          "vehicleId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicleid",
            "description": "UUID of a verified vehicle to update the post to"
          },
          "destinations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Location"
                },
                "type": "array",
                "maxItems": 5,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Destinations",
            "description": "Full replacement list of 1-5 destinations"
          },
          "currentLocation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "contactName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contactname",
            "description": "Override this post's contact name"
          },
          "contactNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contactnumber",
            "description": "Override this post's contact number"
          }
        },
        "type": "object",
        "title": "EditTruckPostRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Location": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable place name"
          },
          "lat": {
            "type": "number",
            "title": "Lat",
            "description": "Latitude"
          },
          "lng": {
            "type": "number",
            "title": "Lng",
            "description": "Longitude"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Location source type",
            "default": "local"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode",
            "description": "Postal code"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City",
            "description": "City name"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "State name"
          }
        },
        "type": "object",
        "required": [
          "name",
          "lat",
          "lng"
        ],
        "title": "Location",
        "example": {
          "city": "Mumbai",
          "lat": 19.076,
          "lng": 72.8777,
          "name": "Mumbai, Maharashtra",
          "pincode": "400001",
          "state": "Maharashtra",
          "type": "local"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from POST /api/auth/verify-otp (`accessToken`)"
      }
    }
  },
...
}